pyproject.toml
[project]
name = "kairos"
version = "0.1.0"
description = "Chief of Staff AI Agent - Voice/Intelligence Loop"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "anthropic>=0.40.0",
    "boto3>=1.35.0",
    "aws-lambda-powertools>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "moto>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "boto3-stubs[sns,ssm]>=1.35.0",
]
cdk = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

